+++ packages/inputs/src/types/control-bind.ts
@@ -14,26 +14,23 @@ export type ControlBindInputAttrs<V, O = V> = {
   onBlur?: VoidFunction;
 };
 
-export type ControlBindPathAttrs<G, O, I = O> = {
-  noCache?: boolean;
-  relatedInputs?: (attrs: {
-    controlValue: O;
-    newControlValue: O;
-    globalValue: G;
-    newGlobalValue: G;
-  }) => G;
-
-  input?: (value: I) => O;
-  output?: (value: O) => I;
-};
-
 export type ControlBindPathFn<V extends ControlValue> = <
   K extends GetAllObjectPaths<V>,
+  I = GetPathObjectType<V, K>,
   O = GetPathObjectType<V, K>,
-  I = O,
 >(
   path: K,
-  attrs?: ControlBindPathAttrs<V, O, I>,
+  attrs?: {
+    noCache?: boolean;
+    relatedInputs?: (attrs: {
+      controlValue: O;
+      newControlValue: O;
+      globalValue: V;
+      newGlobalValue: V;
+    }) => V;
+    input?: (value: I) => O;
+    output?: (value: O) => I;
+  },
 ) => ControlBindInputAttrs<O>;
 
 export type ControlBindMethods<V, RV = NonNullable<V>> = {