+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @under-control/core
 
+## 1.7.14
+
+### Patch Changes
+
+- More infer changes
+
+## 1.7.13
+
+### Patch Changes
+
+- More infer fixes
+
 ## 1.7.12
 
 ### Patch Changes