+++ packages/forms/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @under-control/forms
 
+## 1.7.14
+
+### Patch Changes
+
+- More infer changes
+- Updated dependencies
+  - @under-control/validate@1.7.14
+  - @under-control/inputs@1.7.14
+  - @under-control/core@1.7.14
+
+## 1.7.13
+
+### Patch Changes
+
+- More infer fixes
+- Updated dependencies
+  - @under-control/inputs@1.7.13
+  - @under-control/core@1.7.13
+  - @under-control/validate@1.7.13
+
 ## 1.7.12
 
 ### Patch Changes